added todos and refactored update and startup logic into seperate files
